requirements.txt
Django==5.0.1
django_ckeditor==6.7.0
django_notifications_hq==1.8.3
holidays==0.41
httpx==0.27.0
matplotlib==3.8.2
mysql_connector_repackaged==0.3.1
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.2.0
plotly==5.18.0
PyAutoGUI==0.9.54
python_dateutil==2.8.2
pytz==2023.4
razorpay==1.4.1
Requests==2.31.0
uiautomation==2.0.18
xhtml2pdf==0.2.14
django-mathfilters==1.0.0
mysqlclient==2.2.1
APScheduler==3.10.4
pyinstaller==6.4.0
certifi==2024.2.2
django-user-agents==0.4.0
asgiref==3.7.2
google-api-python-client==2.132.0
google-cloud-storage==2.16.0
qrcode==7.4.2
python-barcode==0.15.1
pyzbar==0.1.9
django-allauth==0.63.3
requests-oauthlib==2.0.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
python-decouple==3.8
setuptools==65.5.0